EOF Optimization - don't spit code sections
Optimize the EOF Functions implementation by not splitting code sections
into separate Bytes object.  Instead the EOF evaluation occurs on a
single container space.  When a function is called the PC (relative to
container start, not section start) is moved.

EOF lacks PC introspection and absolute jumps, so where "PC=0" starts is
not observable to the code.
