+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/results/tracing/vm/VmTraceGenerator.java
@@ -182,11 +182,7 @@ public class VmTraceGenerator {
                     code ->
                         op.setSub(
                             new VmTrace(
-                                currentTraceFrame
-                                    .getMaybeCode()
-                                    .get()
-                                    .getContainerBytes()
-                                    .toHexString())));
+                                currentTraceFrame.getMaybeCode().get().getBytes().toHexString())));
           } else {
             op.setCost(op.getCost());
             op.setSub(null);
@@ -291,11 +287,7 @@ public class VmTraceGenerator {
     // set smart contract code
     if (currentTrace != null && "0x".equals(currentTrace.getCode())) {
       currentTrace.setCode(
-          currentTraceFrame
-              .getMaybeCode()
-              .orElse(CodeV0.EMPTY_CODE)
-              .getContainerBytes()
-              .toHexString());
+          currentTraceFrame.getMaybeCode().orElse(CodeV0.EMPTY_CODE).getBytes().toHexString());
     }
   }
