+++ evm/src/main/java/org/hyperledger/besu/evm/code/CodeSection.java
@@ -18,25 +18,31 @@ package org.hyperledger.besu.evm.code;
 
 import java.util.Objects;
 
-import org.apache.tuweni.bytes.Bytes;
-
 //// java17 convert to record
 public final class CodeSection {
-  final Bytes code;
+
+  final int length;
   final int inputs;
   final int outputs;
   final int maxStackHeight;
+  /** The byte offset from the begining of the container that the section starts at */
+  final int entryPoint;
 
   public CodeSection(
-      final Bytes code, final int inputs, final int outputs, final int maxStackHeight) {
-    this.code = code;
+      final int length,
+      final int inputs,
+      final int outputs,
+      final int maxStackHeight,
+      final int entryPoint) {
+    this.length = length;
     this.inputs = inputs;
     this.outputs = outputs;
     this.maxStackHeight = maxStackHeight;
+    this.entryPoint = entryPoint;
   }
 
-  public Bytes getCode() {
-    return code;
+  public int getLength() {
+    return length;
   }
 
   public int getInputs() {
@@ -51,19 +57,23 @@ public final class CodeSection {
     return maxStackHeight;
   }
 
+  public int getEntryPoint() {
+    return entryPoint;
+  }
+
   @Override
   public boolean equals(final Object o) {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     CodeSection that = (CodeSection) o;
-    return inputs == that.inputs
+    return length == that.length
+        && inputs == that.inputs
         && outputs == that.outputs
-        && maxStackHeight == that.maxStackHeight
-        && code.equals(that.code);
+        && maxStackHeight == that.maxStackHeight;
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(code, inputs, outputs, maxStackHeight);
+    return Objects.hash(length, inputs, outputs, maxStackHeight);
   }
 }
