+++ evm/src/main/java/org/hyperledger/besu/evm/frame/MessageFrame.java
@@ -287,6 +287,7 @@ public class MessageFrame {
     this.stack = new OperandStack(maxStackSize);
     this.returnStack = new ReturnStack();
     returnStack.push(new ReturnStack.ReturnStackItem(0, 0, 0));
+    pc = code.isValid() ? code.getCodeSection(0).getEntryPoint() : 0;
     this.output = Bytes.EMPTY;
     this.returnData = Bytes.EMPTY;
     this.logs = new ArrayList<>();
@@ -377,7 +378,7 @@ public class MessageFrame {
     } else {
       returnStack.push(
           new ReturnStack.ReturnStackItem(section, pc + 2, stack.size() - info.getInputs()));
-      pc = -1; // will be +1ed at end of operations loop
+      pc = info.getEntryPoint() - 1; // will be +1ed at end of operations loop
       this.section = calledSection;
       return null;
     }
