+++ evm/src/main/java/org/hyperledger/besu/evm/operation/JumpFOperation.java
@@ -31,7 +31,7 @@ public class JumpFOperation extends AbstractOperation {
 
   @Override
   public OperationResult execute(final MessageFrame frame, final EVM evm) {
-    final byte[] code = frame.getCode().getCodeBytes(frame.getSection()).toArrayUnsafe();
+    final byte[] code = frame.getCode().getBytes().toArrayUnsafe();
     return staticOperation(frame, code, frame.getPC());
   }
