+++ evm/src/main/java/org/hyperledger/besu/evm/tracing/StandardJsonTracer.java
@@ -66,7 +66,9 @@ public class StandardJsonTracer implements OperationTracer {
     for (int i = messageFrame.stackSize() - 1; i >= 0; i--) {
       stack.add("\"" + shortBytes(messageFrame.getStackItem(i)) + "\"");
     }
-    pc = messageFrame.getPC();
+    pc =
+        messageFrame.getPC()
+            - messageFrame.getCode().getCodeSection(messageFrame.getSection()).getEntryPoint();
     section = messageFrame.getSection();
     gas = shortNumber(messageFrame.getRemainingGas());
     memorySize = messageFrame.memoryWordSize() * 32;
